Essensular Rod
Magicks! This (yet unobtainable) rod will allow you to insert 1-2
essences to harness their power. Whether you use it as a weapon, a
utility or a little bit of both is up to you. Will take up to 512 uses
before breaking, and will use up one of the 16 uses of each essence in
the process, too.

Other things, also.